clean up/restructure code for future enhancements
