+++ extensions/mistral/requirements.txt
@@ -1,2 +1,1 @@
-MistralAI
-Mistral
+mistralai
